split search command
